requirements.txt
Django>=4.2.17
djangorestframework==3.14.0
django-cors-headers==4.3.0
pyotp==2.9.0
python-dotenv==1.0.0